pyproject.toml
[project]
name = "neurogated"
version = "0.1.0"
description = "Neuro-Gated Graph Memory System - A neurobiologically inspired retrieval mechanism"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "openai>=1.0.0",
    "python-igraph>=0.11.0",
    "pandas>=2.0.0",
    "tqdm>=4.65.0",
    "tiktoken>=0.7.0",
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"
